requirements.txt
matplotlib
pandas
yfinance
alpaca_py